move config and update package.json
