+++ package.json
@@ -15,13 +15,14 @@
     "lint": "eslint . --max-warnings 0",
     "lint:fix": "eslint . --fix",
     "dev": "npx run build && tsx dist/esm/src/index.js",
-    "format": "npx prettier --write ./src/**/**/**/*"
+    "format": "npx prettier --write ./src/**/**/**/*",
+    "workflow": "npm ci && npm run build --if-present && npm test"
   },
-  "homepage": "https://github.com/TBD54566975/incubation-tblend#readme",
-  "bugs": "https://github.com/TBD54566975/incubation-tblend/issues",
+  "homepage": "https://github.com/TBD54566975/incubation-dcx#readme",
+  "bugs": "https://github.com/TBD54566975/incubation-dcx/issues",
   "repository": {
     "type": "git",
-    "url": "git+ssh://git@github.com/TBD54566975/incubation-tblend.git"
+    "url": "git+ssh://git@github.com/TBD54566975/incubation-dcx.git"
   },
   "license": "Apache-2.0",
   "contributors": [
@@ -41,13 +42,7 @@
   "exports": {
     ".": {
       "types": "./dist/types/index.d.ts",
-      "import": "./dist/esm/index.js",
-      "require": "./dist/cjs/index.js"
-    },
-    "./utils": {
-      "types": "./dist/types/utils.d.ts",
-      "import": "./dist/esm/utils.js",
-      "require": "./dist/cjs/utils.js"
+      "import": "./dist/esm/index.js"
     }
   },
   "keywords": [
