+++ src/core/index.ts
@@ -1,1 +1,6 @@
 export * from './web5-manager.js';
+export * from './config.js';
+export * from './dcx-server.js';
+
+import server from './dcx-server.js';
+export default server;
