+++ src/types/index.ts
@@ -1,3 +1,4 @@
 export type * from './cipher.js';
 export type * from './dcx.js';
+export type * from './did.js';
 export type * from './web5.js';
