+++ src/utils/index.ts
@@ -4,5 +4,5 @@ export * from './error.js';
 export * from './file-system.js';
 export * from './json.js';
 export * from './logger.js';
-export * from './time.js';
 export * from './objects.js';
+export * from './time.js';
